Fix Chrome rendering bug
by including space in our set of glyphs. Note that I also removed the
U+0331 character from text and added it to a unicode list instead. It's
a combining character, so it's difficult to work with in strings. I'm
not totally sure why we have it, but I'm also anxious to get rid of it.

Ref: https://www.fileformat.info/info/unicode/char/0331/index.htm